+++ packages/article-preview/src/App.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Card } from './components';
 
 import image from './images/drawers.jpg';