+++ packages/social-media-dashboard/src/app.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styled from 'styled-components';
 
 import { AppProvider } from 'app-context';