+++ packages/social-media-dashboard/src/components/dashboard-header/dashboard-header.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import React, { useContext } from 'react';
+import { useContext } from 'react';
 
 import { mediaQuery } from 'global-style';
 import { AppStateContext } from 'app-context';