+++ packages/social-media-dashboard/src/components/subheading/subheading.spec.tsx
@@ -1,10 +1,8 @@
-import React from 'react';
-
 import { RenderResult } from '@testing-library/react';
 
 import { renderAppProvider, renderWithTheme } from '__mocks__/app-context-mock';
-import { Subheading } from './subheading';
 import { themes } from 'models';
+import { Subheading } from './subheading';
 
 describe('<Subheading />', () => {
   let rendered: RenderResult;