+++ packages/social-media-dashboard/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/ezavile/conquer-react/compare/@conquer-react/social-media-dashboard@0.2.0...@conquer-react/social-media-dashboard@0.3.0) (2021-02-09)
+
+
+### Features
+
+* **social-media-dashboard:** adjustments after upgrading to React@17 ([0e10e80](https://github.com/ezavile/conquer-react/commit/0e10e8050b7d598c1aa5cb5b5a52e512b7dceabb))
+
+
+
+
+
 # [0.2.0](https://github.com/ezavile/conquer-react/compare/@conquer-react/social-media-dashboard@0.1.0...@conquer-react/social-media-dashboard@0.2.0) (2020-11-10)